Adjust OpenStack / POWER CI Access drop down
This provides a drop down that asks users if they need only OpenStack or POWER
CI or both to help with our on boarding process.

In addition, I move some of the questions around and grouped them so that people
know that one section is for CI vs. OpenStack.

Ideally, it would be nice if we had Javascript which automatically displayed the
second they need to fill out, but my JS-fu is lacking in that.